fix: adding  explict cast

+++ src/wrapping_integers.cc
@@ -6,7 +6,7 @@ using namespace std;
 
 Wrap32 Wrap32::wrap( uint64_t n, Wrap32 zero_point )
 {
-  return zero_point + n;
+  return zero_point + static_cast<uint32_t>(n);
 }
 
 uint64_t Wrap32::unwrap( Wrap32 zero_point, uint64_t checkpoint ) const
@@ -16,7 +16,7 @@ uint64_t Wrap32::unwrap( Wrap32 zero_point, uint64_t checkpoint ) const
    * n = (zero_point - raw_data_) + k * 2^32 = offset + k * 2^32
    * Since we need to find the corresponding absolute sequnence number that is closest to checkponit,
    * so consider base of checkpoint as rough base. When the absolute sequence number is in the same base,
-   *  rough base is the real base. Otherwise check the nearby by shifting rough base.
+   * rough base is the real base. Otherwise check the nearby by shifting rough base.
    */
   uint64_t offset = raw_value_ - zero_point.raw_value_;
   uint64_t rough_base = checkpoint & ~0xFFFFFFFFULL;
